誰がログインしているかを判定する機能まで

+++ app/api/item/delete/[id]/route.ts
@@ -5,12 +5,24 @@ export async function DELETE(
   request: NextRequest,
   { params }: { params: Promise<{ [key: string]: string | string[] }> }
 ) {
+  const reqBody = await request.json();
   const { id } = await params;
 
   try {
-    const { error } = await supabase.from("items").delete().eq("id", id);
+    const { data, error } = await supabase
+      .from("items")
+      .select()
+      .eq("id", id)
+      .single();
     if (error) throw new Error(error.message);
-    return NextResponse.json({ message: "アイテム削除成功" });
+
+    if (data.email === reqBody.email) {
+      const { error } = await supabase.from("items").delete().eq("id", id);
+      if (error) throw new Error(error.message);
+      return NextResponse.json({ message: "アイテム削除成功" });
+    } else {
+      return NextResponse.json({ message: "他の人が作成したアイテムです" });
+    }
   } catch (err) {
     return NextResponse.json({ message: `アイテム削除失敗：${err}` });
   }
